Refactored luminary file functions.
